+++ src/contracts/evm/Account/abi.json
@@ -1,4 +1,6 @@
 [
+  "event EndowmentCreated(uint256,tuple(address,string,tuple(uint256[],uint256[]),uint256,uint8,string,string,uint8,bool,bool,uint256,tuple(string[],uint256[],string[],uint256[]),tuple(string[],uint256[],string[],uint256[]),tuple(bool,bool,uint256,bool,uint256),bool,uint256,uint256,uint256,address,address,bool,address,address[],address[],address[],tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256))),uint256,bool,tuple(uint256,uint256,uint256)))",
+  "event UpdateEndowmentState(uint256,tuple(tuple(uint256,uint256),tuple(tuple(uint256,uint256[],address[]),tuple(uint256,uint256[],address[])),bool,tuple(tuple(uint256,address),uint8)))",
   "function createEndowment(tuple(address,bool,uint256,uint256,string,tuple(uint256[],uint256[]),uint256,uint8,string,string,address[],bool,tuple(uint8,tuple(uint256,uint256,uint256,uint256)),tuple(uint8,tuple(uint256,uint256)),address[],address[],uint256,uint256,uint256,tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(uint256,uint256,uint256,uint256,uint256,uint128,uint256,tuple(uint8,tuple(address,uint256,string,string,tuple(uint8,tuple(uint128,uint256,uint128,uint128)),string,string,uint256,address,uint256,uint256))),bool,uint256,tuple(tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256))),uint256,address[],bool,tuple(uint256,uint256,uint256))) returns (uint256)",
-  "event EndowmentCreated(uint256,tuple(address,string,tuple(uint256[],uint256[]),uint256,uint8,string,string,uint8,bool,bool,uint256,tuple(string[],uint256[],string[],uint256[]),tuple(string[],uint256[],string[],uint256[]),tuple(bool,bool,uint256,bool,uint256),bool,uint256,uint256,uint256,address,address,bool,address,address[],address[],address[],tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(address,uint256,bool),tuple(tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256)),tuple(bool,bool,bool,tuple(address,uint256))),uint256,bool,tuple(uint256,uint256,uint256)))"
+  "function withdraw(uint256,uint8,address,address[],uint256[])"
 ]
