+++ src/contracts/evm/index.ts
@@ -1,3 +1,3 @@
-export * as AccountDepositWithdrawEndowments from "./AccountDepositWithdrawEndowments";
+export * as Account from "./Account";
 export * as EndowmentMultiSig from "./EndowmentMultiSig";
 export * as LockedWithdraw from "./LockedWithdraw";
